add parameter to read bin file

+++ inc/bus/bus.hpp
@@ -2,8 +2,8 @@
 #include <cstdint>
 
 struct Bus {
-    uint16_t address = 0;
-    uint16_t data = 0;
+    uint32_t address = 0;
+    uint32_t data = 0;
     bool read = false;
     bool write = false;
 };
